backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
pydantic[email]==2.5.0

# Production & Scaling Dependencies
redis==5.0.1
prometheus-client==0.19.0
slowapi==0.1.9
structlog==23.2.0

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2